+++ contracts/src/core/ERC20Mock.sol
@@ -3,8 +3,8 @@
 
 pragma solidity ^0.8.25;
 
-import "@openzeppelin/contracts/interfaces/IERC20.sol";
-import "@openzeppelin/contracts/utils/Context.sol";
+import '@openzeppelin/contracts/interfaces/IERC20.sol';
+import '@openzeppelin/contracts/utils/Context.sol';
 
 /**
  * @dev Implementation of the {IERC20} interface.
@@ -35,29 +35,27 @@ import "@openzeppelin/contracts/utils/Context.sol";
  * allowances. See {IERC20-approve}.
  */
 contract ERC20Mock is Context, IERC20 {
-    mapping(address => uint256) private _balances;
+    mapping(address => uint) private _balances;
 
-    mapping(address => mapping(address => uint256)) private _allowances;
+    mapping(address => mapping(address => uint)) private _allowances;
 
-    uint256 private _totalSupply;
+    uint private _totalSupply;
 
     /**
      * @dev See {IERC20-totalSupply}.
      */
-    function totalSupply() public view virtual override returns (uint256) {
+    function totalSupply() public view virtual override returns (uint) {
         return _totalSupply;
     }
 
     /**
      * @dev See {IERC20-balanceOf}.
      */
-    function balanceOf(
-        address account
-    ) public view virtual override returns (uint256) {
+    function balanceOf(address account) public view virtual override returns (uint) {
         return _balances[account];
     }
 
-    function mint(address account, uint256 amount) public {
+    function mint(address account, uint amount) public {
         _mint(account, amount);
     }
 
@@ -69,10 +67,7 @@ contract ERC20Mock is Context, IERC20 {
      * - `to` cannot be the zero address.
      * - the caller must have a balance of at least `amount`.
      */
-    function transfer(
-        address to,
-        uint256 amount
-    ) public virtual override returns (bool) {
+    function transfer(address to, uint amount) public virtual override returns (bool) {
         address owner = _msgSender();
         _transfer(owner, to, amount);
         return true;
@@ -81,10 +76,7 @@ contract ERC20Mock is Context, IERC20 {
     /**
      * @dev See {IERC20-allowance}.
      */
-    function allowance(
-        address owner,
-        address spender
-    ) public view virtual override returns (uint256) {
+    function allowance(address owner, address spender) public view virtual override returns (uint) {
         return _allowances[owner][spender];
     }
 
@@ -98,10 +90,7 @@ contract ERC20Mock is Context, IERC20 {
      *
      * - `spender` cannot be the zero address.
      */
-    function approve(
-        address /*spender*/,
-        uint256 /*amount*/
-    ) public virtual override returns (bool) {
+    function approve(address, /*spender*/ uint /*amount*/ ) public virtual override returns (bool) {
         return true;
     }
 
@@ -121,11 +110,7 @@ contract ERC20Mock is Context, IERC20 {
      * - the caller must have allowance for ``from``'s tokens of at least
      * `amount`.
      */
-    function transferFrom(
-        address from,
-        address to,
-        uint256 amount
-    ) public virtual override returns (bool) {
+    function transferFrom(address from, address to, uint amount) public virtual override returns (bool) {
         _transfer(from, to, amount);
         return true;
     }
@@ -144,20 +129,13 @@ contract ERC20Mock is Context, IERC20 {
      * - `to` cannot be the zero address.
      * - `from` must have a balance of at least `amount`.
      */
-    function _transfer(
-        address from,
-        address to,
-        uint256 amount
-    ) internal virtual {
-        require(from != address(0), "ERC20: transfer from the zero address");
-        require(to != address(0), "ERC20: transfer to the zero address");
+    function _transfer(address from, address to, uint amount) internal virtual {
+        require(from != address(0), 'ERC20: transfer from the zero address');
+        require(to != address(0), 'ERC20: transfer to the zero address');
 
         _beforeTokenTransfer(from, to, amount);
 
-        require(
-            _balances[from] >= amount,
-            "ERC20: transfer amount exceeds balance"
-        );
+        require(_balances[from] >= amount, 'ERC20: transfer amount exceeds balance');
         unchecked {
             _balances[from] = _balances[from] - amount;
             // Overflow not possible: the sum of all balances is capped by totalSupply, and the sum is preserved by
@@ -170,7 +148,8 @@ contract ERC20Mock is Context, IERC20 {
         _afterTokenTransfer(from, to, amount);
     }
 
-    /** @dev Creates `amount` tokens and assigns them to `account`, increasing
+    /**
+     * @dev Creates `amount` tokens and assigns them to `account`, increasing
      * the total supply.
      *
      * Emits a {Transfer} event with `from` set to the zero address.
@@ -179,8 +158,8 @@ contract ERC20Mock is Context, IERC20 {
      *
      * - `account` cannot be the zero address.
      */
-    function _mint(address account, uint256 amount) internal virtual {
-        require(account != address(0), "ERC20: mint to the zero address");
+    function _mint(address account, uint amount) internal virtual {
+        require(account != address(0), 'ERC20: mint to the zero address');
 
         _totalSupply += amount;
         unchecked {
@@ -201,13 +180,13 @@ contract ERC20Mock is Context, IERC20 {
      * - `account` cannot be the zero address.
      * - `account` must have at least `amount` tokens.
      */
-    function _burn(address account, uint256 amount) internal virtual {
-        require(account != address(0), "ERC20: burn from the zero address");
+    function _burn(address account, uint amount) internal virtual {
+        require(account != address(0), 'ERC20: burn from the zero address');
 
         _beforeTokenTransfer(account, address(0), amount);
 
-        uint256 accountBalance = _balances[account];
-        require(accountBalance >= amount, "ERC20: burn amount exceeds balance");
+        uint accountBalance = _balances[account];
+        require(accountBalance >= amount, 'ERC20: burn amount exceeds balance');
         unchecked {
             _balances[account] = accountBalance - amount;
             // Overflow not possible: amount <= accountBalance <= totalSupply.
@@ -232,13 +211,9 @@ contract ERC20Mock is Context, IERC20 {
      * - `owner` cannot be the zero address.
      * - `spender` cannot be the zero address.
      */
-    function _approve(
-        address owner,
-        address spender,
-        uint256 amount
-    ) internal virtual {
-        require(owner != address(0), "ERC20: approve from the zero address");
-        require(spender != address(0), "ERC20: approve to the zero address");
+    function _approve(address owner, address spender, uint amount) internal virtual {
+        require(owner != address(0), 'ERC20: approve from the zero address');
+        require(spender != address(0), 'ERC20: approve to the zero address');
 
         _allowances[owner][spender] = amount;
         emit Approval(owner, spender, amount);
@@ -252,17 +227,10 @@ contract ERC20Mock is Context, IERC20 {
      *
      * Might emit an {Approval} event.
      */
-    function _spendAllowance(
-        address owner,
-        address spender,
-        uint256 amount
-    ) internal virtual {
-        uint256 currentAllowance = allowance(owner, spender);
-        if (currentAllowance != type(uint256).max) {
-            require(
-                currentAllowance >= amount,
-                "ERC20: insufficient allowance"
-            );
+    function _spendAllowance(address owner, address spender, uint amount) internal virtual {
+        uint currentAllowance = allowance(owner, spender);
+        if (currentAllowance != type(uint).max) {
+            require(currentAllowance >= amount, 'ERC20: insufficient allowance');
         }
     }
 
@@ -280,11 +248,7 @@ contract ERC20Mock is Context, IERC20 {
      *
      * To learn more about hooks, head to xref:ROOT:extending-contracts.adoc#using-hooks[Using Hooks].
      */
-    function _beforeTokenTransfer(
-        address from,
-        address to,
-        uint256 amount
-    ) internal virtual {}
+    function _beforeTokenTransfer(address from, address to, uint amount) internal virtual { }
 
     /**
      * @dev Hook that is called after any transfer of tokens. This includes
@@ -300,9 +264,5 @@ contract ERC20Mock is Context, IERC20 {
      *
      * To learn more about hooks, head to xref:ROOT:extending-contracts.adoc#using-hooks[Using Hooks].
      */
-    function _afterTokenTransfer(
-        address from,
-        address to,
-        uint256 amount
-    ) internal virtual {}
+    function _afterTokenTransfer(address from, address to, uint amount) internal virtual { }
 }
